+++ src/godot_pool_byte_array.lua
@@ -258,7 +258,7 @@ PoolByteArray = ffi_metatype('godot_pool_byte_array', {
 		methods.push_back(self, ...)
 		return self
 	end,
-	__gc = godot_pool_byte_array_destroy,
+	__gc = api.godot_pool_byte_array_destroy,
 	--- Returns method named `index` or the result of `safe_get`.
 	-- @function __index
 	-- @param index