+++ src/godot_pool_string_array.lua
@@ -240,7 +240,7 @@ PoolStringArray = ffi_metatype('godot_pool_string_array', {
 		methods.push_back(self, ...)
 		return self
 	end,
-	__gc = godot_pool_string_array_destroy,
+	__gc = api.godot_pool_string_array_destroy,
 	--- Returns method named `index` or the result of `safe_get`.
 	-- @function __index
 	-- @param index